approot/requirements.txt
flask
databricks-sql-connector[pyarrow]
python-dotenv
requests
